+++ fbw-common/src/typings/fs-base-ui/html_ui/JS/Avionics.d.ts
@@ -1,4 +1,4 @@
-/// <reference path="../../../units.d.ts" />
+/// <reference path="../../../types.d.ts" />
 /// <reference path="./Types.d.ts" />
 
 declare global {
