+++ fbw-common/src/typings/fs-base-ui/html_ui/JS/simvar.d.ts
@@ -1,4 +1,4 @@
-/// <reference path="../../../units.d.ts" />
+/// <reference path="../../../types.d.ts" />
 /// <reference path="./Types.d.ts" />
 /// <reference path="./Simplane.d.ts" />
