+++ fbw-common/src/typings/index.d.ts
@@ -3,5 +3,5 @@ export * from './asobo-vcockpits-instruments-a320-neo';
 export * from './asobo-vcockpits-instruments-airliners';
 export * from './flybywire-vcockpits-instruments';
 export * from './fs-base-ui';
+export * from './fmgc';
 export * from './types';
-export * from './units';
