--- fbw-common/src/typings/units.d.ts
@@ -1,34 +1,0 @@
-declare global {
-    type NauticalMiles = number;
-    type Heading = number;
-    type Track = number;
-    type Latitude = number;
-    type Longitude = number;
-    type Feet = number;
-    type FlightLevel = number;
-    type Knots = number;
-    type FeetPerMinute = number;
-    type Metres = number;
-    type MetresPerSecond = number;
-    type Mach = number;
-    type Degrees = number;
-    type DegreesMagnetic = number;
-    type DegreesTrue = number;
-    type Seconds = number;
-    type Minutes = number;
-    type Percent = number;
-    type Radians = number;
-    type RotationsPerMinute = number;
-    type Angl16 = number;
-    type RadiansPerSecond = number;
-    type PercentOver100 = number;
-    type Gallons = number;
-    type Kilograms = number;
-    type Pounds = number;
-    type Celsius = number;
-    type InchesOfMercury = number;
-    type Millibar = number;
-    type PressurePerSquareInch = number;
-}
-
-export {};
